backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
PyJWT==2.9.0
passlib==1.7.4
werkzeug==3.0.4
pytesseract==0.3.13
pdf2image==1.17.0
Pillow==10.4.0
numpy==1.26.4
PyMuPDF==1.26.5
opencv-python==4.6.0.66
rapidfuzz==3.14.1
pandas==2.3.3
